Remove dummy list of questions from testSlice and use question data passed from the Question component to form user results.
This means that all the data on the questions themselves is coming from the API via RTK Query. But note that I haven't figured out a clean way of dynamically getting the total number of questions into the local testSlice state. This was previously just the length of the dummy list array. Now it's hard coded as 4, since that's the total I have while testing. When the app is complete, it's possible to keep it hard coded at 30, since that's what we plan to use for all students. But it would be better to dynamically get it.

Right now, it's not possible from startTest reducer, because that is called from the start page, before any API queries are made. But perhaps in the future, when the user will have to enter some info to be sent to the API, it could be possible.